sumal edit 2025.07.24

+++ Model/EventType.cs
@@ -1,6 +1,6 @@
 public class EventType
 {
-    public int EventTypeID { get; set; }
+    public long EventTypeID { get; set; }
     public string EventCode { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public string? Remarks { get; set; }
