+++ Model/RoomType.cs
@@ -1,6 +1,6 @@
 public class RoomType
 {
-    public int RoomTypeID { get; set; }
+    public long RoomTypeID { get; set; }
     public string RoomTypeCode { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public string? Remarks { get; set; }
